set deployment configs

+++ app/src/main/java/com/imquarantined/ui/service/EndlessService.kt
@@ -287,7 +287,7 @@ class EndlessService : Service(), CustomServiceTask {
                 launch(Dispatchers.IO) {
                     backGroundWork()
                 }
-                delay(Const.Config.backgroundTaskPeriod)
+                delay(Const.Config.backgroundTaskPeriod.toLong())
             }
             Timber.i("End of the loop for the service")
         }
